# frontend/openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MonthViewResponse": {
        "properties": {
          "days": {
            "items": {
              "$ref": "#/components/schemas/WeekDayGroupResponse"
            },
            "title": "Days",
            "type": "array"
          },
          "month": {
            "title": "Month",
            "type": "string"
          },
          "month_end": {
            "format": "date",
            "title": "Month End",
            "type": "string"
          },
          "month_start": {
            "format": "date",
            "title": "Month Start",
            "type": "string"
          },
          "week_start_day": {
            "title": "Week Start Day",
            "type": "string"
          }
        },
        "required": ["month", "month_start", "month_end", "week_start_day", "days"],
        "title": "MonthViewResponse",
        "type": "object"
      },
...
      "TodayItemResponse": {
        "properties": {
          "due_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "ends_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          },
          "is_all_day": {
            "default": false,
            "title": "Is All Day",
            "type": "boolean"
          },
          "is_completed": {
            "default": false,
            "title": "Is Completed",
            "type": "boolean"
          },
          "is_overdue": {
            "title": "Is Overdue",
            "type": "boolean"
          },
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ViewItemKind"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "occurrence_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurrence Role"
          },
          "routine_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routine Title"
          },
          "span_end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Span End Date"
          },
          "span_segment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Span Segment"
          },
          "span_start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Span Start Date"
          },
          "starts_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts At"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "item_id",
          "title",
          "notes",
          "due_date",
          "starts_at",
          "ends_at",
          "is_overdue"
        ],
        "title": "TodayItemResponse",
        "type": "object"
      },
      "TodayViewResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TodayItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "reference_date": {
            "format": "date",
            "title": "Reference Date",
            "type": "string"
          }
        },
        "required": ["reference_date", "items"],
        "title": "TodayViewResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      },
      "ViewItemKind": {
        "enum": ["task", "occurrence", "appointment", "maintenance"],
        "title": "ViewItemKind",
        "type": "string"
      },
      "WeekDayGroupResponse": {
        "properties": {
          "date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/WeekItemResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          }
        },
        "required": ["date", "items"],
        "title": "WeekDayGroupResponse",
        "type": "object"
      },
      "WeekItemResponse": {
        "properties": {
          "due_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "ends_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          },
          "is_all_day": {
            "default": false,
            "title": "Is All Day",
            "type": "boolean"
          },
          "is_completed": {
            "default": false,
            "title": "Is Completed",
            "type": "boolean"
          },
          "is_overdue": {
            "title": "Is Overdue",
            "type": "boolean"
          },
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ViewItemKind"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "occurrence_role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurrence Role"
          },
          "routine_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Routine Title"
          },
          "span_end_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Span End Date"
          },
          "span_segment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Span Segment"
          },
          "span_start_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Span Start Date"
          },
          "starts_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Starts At"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "item_id",
          "title",
          "due_date",
          "starts_at",
          "ends_at",
          "is_overdue"
        ],
        "title": "WeekItemResponse",
        "type": "object"
      },
      "WeekTargetSummaryResponse": {
        "properties": {
          "completed_count": {
            "title": "Completed Count",
            "type": "integer"
          },
          "target_count": {
            "title": "Target Count",
            "type": "integer"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": ["target_id", "title", "completed_count", "target_count"],
        "title": "WeekTargetSummaryResponse",
        "type": "object"
      },
      "WeekViewResponse": {
        "properties": {
          "days": {
            "items": {
              "$ref": "#/components/schemas/WeekDayGroupResponse"
            },
            "title": "Days",
            "type": "array"
          },
          "targets": {
            "items": {
              "$ref": "#/components/schemas/WeekTargetSummaryResponse"
            },
            "title": "Targets",
            "type": "array"
          },
          "week_end": {
            "format": "date",
            "title": "Week End",
            "type": "string"
          },
          "week_start": {
            "format": "date",
            "title": "Week Start",
            "type": "string"
          }
        },
        "required": ["week_start", "week_end", "days", "targets"],
        "title": "WeekViewResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Self-hosted personal planning platform API",
    "title": "Planforge",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/views/month": {
      "get": {
        "description": "Return the Month view for a YYYY-MM calendar month.",
        "operationId": "month_view_endpoint_api_views_month_get",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonthViewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Month View Endpoint",
        "tags": ["views"]
      }
    },
    "/api/views/today": {
      "get": {
        "description": "Return the Today view for a reference date.",
        "operationId": "today_view_endpoint_api_views_today_get",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodayViewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Today View Endpoint",
        "tags": ["views"]
      }
    },
    "/api/views/week": {
      "get": {
        "description": "Return the Week view for a week starting on the given date.",
        "operationId": "week_view_endpoint_api_views_week_get",
        "parameters": [
          {
            "in": "query",
            "name": "week_start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Week Start"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeekViewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Week View Endpoint",
        "tags": ["views"]
      }
    },
...
  }
}